Hotfix v0.4.1: Market UI arithmetic decoupling, Energy grid clamping, and World-Gen dynamic scaling
Bugfix and UI/UX sprint addressing five anomaly groups:

1. Market Net Surplus discrepancy:
   - Added net_trade and b2c_demand_volume fields to GlobalMarket
   - Added CommodityTradeEntry to TradeDelta for per-commodity trade flows
   - Clear/repopulate supply_volume, demand_volume, b2c_demand_volume, net_trade
     at turn start before B2B aggregation
   - Aggregate net_trade after balance_global_trade
   - Backend market.net_surplus preserved as raw B2B clearing signal (untouched)
   - UI net_surplus computed dynamically in snapshot.rs as
     supply - demand + net_trade (temporal-causality fix)

2. Market UI refactor:
   - Removed Last Trade column from CommodityRow DTO and MarketPage
   - Added Net Trade column with import/export color coding
   - Status column gated behind Show Inactive toggle (hidden by default)

3. Energy region ID leakage:
   - Region display uses display_name instead of backend id
   - Internal region_id retained for keys and data lookups

4. Static world-generation scaling:
   - Use computed plant_count (was discarded into _plant_count)
   - Replace hardcoded 500.0 wage fallback with average_wage.max(1.0)
   - Feed real geographic traits into plant selection
   - Scale plant counts, nameplate capacity, and workers with country data

5. Energy Grid math and logic:
   - Clamp generated supply to nameplate capacity (no matter from void)
   - Add effective_supply_mw to RegionEnergyInfo DTO
   - Display effective supply in EnergyPage frontend
   - Derive LV capacity from actual connected housing/commercial electricity
     demand (replaces pop*dev*wage/500_000 magic formula)
   - Move grid initialization after generate_corporate_entities
   - Remove employment-based fallback nameplate capacity estimation
   - Only buildings with PowerPlantMetadata count toward max_capacity_mw

Tests:
- market_identity_test.rs: 8 tests (UI identity, B2C separation, net_trade)
- world_gen_scaling_test.rs: 5 tests (plant count, wage floor, scaling)
- energy_wave1_test.rs: 6 new tests (supply clamping, effective supply,
  load-shed logic, region display name)

CI/CD:
- Smart stop hook bypasses pipeline for documentation/config-only changes
- 2051 tests pass, 0 failures
- npm run build succeeds

+++ state/src/corporate/bounded_rationality.rs
@@ -3,9 +3,7 @@
 //! This module implements information access tiers and market research mechanics
 //! that simulate realistic decision-making limitations for companies of different sizes.
 
-use crate::registries::enums::{Commodity, Sector};
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
 
 /// Information quality tier for corporate decision-making.
 ///
@@ -114,7 +112,7 @@ pub fn apply_estimation_error(true_cost: f64, quality: InformationQuality) -> f6
     }
     // Use a deterministic hash of the cost to produce a stable pseudo-random
     // deviation in [-error_rate, +error_rate] without requiring an RNG.
-    let hash = (true_cost.to_bits().wrapping_add(0x9E3779B97F4A7C15) as u64) as f64;
+    let hash = true_cost.to_bits().wrapping_add(0x9E3779B97F4A7C15) as f64;
     let frac = (hash / u64::MAX as f64) * 2.0 - 1.0; // [-1, 1]
     true_cost * (1.0 + frac * error_rate)
 }
